Refactored trade commands
